Refactor Footer component into component file Footer
